Add authRouter to another file not indexRouter
