+++ app.js
@@ -11,6 +11,7 @@ const compression = require('compression');
 const logger = require('morgan');
 
 const indexRouters = require('./routes/index');
+const authRouter = require('./routes/auth');
 const directRouters = require('./routes/direct');
 const apiRouters = require('./routes/api');
 
@@ -90,6 +91,7 @@ app.get('/docs', isAuthenticated, async (req, res) => {
 });
 
 app.use('/', indexRouters);
+app.use('/auth', authRouter)
 app.use('/api', apiRouters);
 app.use('/direct', directRouters);
