+++ views/index.ejs
@@ -6,7 +6,7 @@
       Let's make this world more colorful with programming language
     </p>
     <div class="d-grid gap-2 d-sm-flex justify-content-sm-center">
-      <a class="btn btn-outline-gradient py-2" href="/login"
+      <a class="btn btn-outline-gradient py-2" href="/auth/login"
         ><span>Sign in</span></a
       >
     </div>
