+++ views/login.ejs
@@ -19,7 +19,7 @@
 		<h1 class="h3 mb-3 fw-normal">Sign In</h1>
 		<%- include('layouts/msg')  %> 
         <%- include('layouts/errors')  %> 
-      <form method="post" action="/login">
+      <form method="post" action="/auth/login">
         <div class="form-floating">
           <input type="text" name="username" class="form-control" id="floatingInput" required/>
           <label for="floatingInput">Username</label>
@@ -33,7 +33,7 @@
         </div>
         <div class="checkbox mb-3">
           <label>
-            Don't have an account? <a href="/register">register</a>
+            Don't have an account? <a href="/auth/register">register</a>
           </label>
         </div>
         <button class="w-100 btn btn-lg btn-primary" type="submit">
